Also copy adjustments and additional charges
